XY shifter support

+++ brWheel_my/ConfigHID.h
@@ -8,7 +8,7 @@ typedef struct
 {
   uint8_t ReportId;
   uint16_t Rotation;
-  int16_t Offset; //milos, added
+  int32_t Offset; //milos, added
   uint8_t GeneralGain;
   uint8_t ConstantGain;
   uint8_t DamperGain;
